+++ packages/ApiTests/src/package-test.ts
@@ -48,6 +48,7 @@ import './dapi/domain-search';
 import './dapi/domain-visual-queries';
 import './dapi/domain-external-write';
 import './dapi/domain-external';
+import './dapi/domain-authoring';
 import './dapi/entity-properties';
 import './dapi/connector-writes';
 import './dapi/connector-ddl';